requirements.txt
-c constraints.txt

requests_oauthlib
Django
drf-spectacular
djangorestframework
psycopg2
djangorestframework-simplejwt

# test app
Flask
